Program/requirements.txt
yt-dlp==2024.12.23
pydub==0.25.1
openai-whisper==20231117
torch
torchaudio
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.5
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]
email-validator
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
sqlmodel==0.0.14
asyncpg==0.29.0
greenlet==3.0.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
google-api-python-client==2.115.0
